# apps/api/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ActivateWorkflowRolloutRequest": {
        "properties": {
          "candidate_version_id": {
            "format": "uuid",
            "title": "Candidate Version Id",
            "type": "string"
          }
        },
        "required": [
          "candidate_version_id"
        ],
        "title": "ActivateWorkflowRolloutRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublishWorkflowVersionRequest": {
        "properties": {
          "published_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published By"
          }
        },
        "title": "PublishWorkflowVersionRequest",
        "type": "object"
      },
...
      "WorkflowCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "domain": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Domain",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "domain"
        ],
        "title": "WorkflowCreate",
        "type": "object"
      },
...
      "WorkflowRead": {
        "properties": {
          "active_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "slug",
          "name",
          "domain",
          "created_at"
        ],
        "title": "WorkflowRead",
        "type": "object"
      },
...
      "WorkflowVersionCreate": {
        "properties": {
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "eval_dataset_bindings": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Eval Dataset Bindings",
            "type": "array"
          },
          "guardrails": {
            "items": {
              "type": "string"
            },
            "title": "Guardrails",
            "type": "array"
          },
          "input_schema": {
            "additionalProperties": true,
            "title": "Input Schema",
            "type": "object"
          },
          "model_config": {
            "$ref": "#/components/schemas/ModelConfig"
          },
          "output_schema": {
            "additionalProperties": true,
            "title": "Output Schema",
            "type": "object"
          },
          "policy_pack": {
            "items": {
              "$ref": "#/components/schemas/WorkflowPolicyRule"
            },
            "title": "Policy Pack",
            "type": "array"
          },
          "prompt_template": {
            "minLength": 1,
            "title": "Prompt Template",
            "type": "string"
          },
          "rollout_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowRolloutConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "tool_set": {
            "items": {
              "$ref": "#/components/schemas/WorkflowToolRef"
            },
            "title": "Tool Set",
            "type": "array"
          },
          "version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "prompt_template",
          "input_schema",
          "output_schema",
          "model_config"
        ],
        "title": "WorkflowVersionCreate",
        "type": "object"
      },
      "WorkflowVersionRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "eval_dataset_bindings": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Eval Dataset Bindings",
            "type": "array"
          },
          "guardrails": {
            "items": {
              "type": "string"
            },
            "title": "Guardrails",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_schema": {
            "additionalProperties": true,
            "title": "Input Schema",
            "type": "object"
          },
          "model_config": {
            "$ref": "#/components/schemas/ModelConfig"
          },
          "output_schema": {
            "additionalProperties": true,
            "title": "Output Schema",
            "type": "object"
          },
          "policy_pack": {
            "items": {
              "$ref": "#/components/schemas/WorkflowPolicyRule"
            },
            "title": "Policy Pack",
            "type": "array"
          },
          "prompt_template": {
            "title": "Prompt Template",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "rollout_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowRolloutConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowVersionStatus"
          },
          "tool_set": {
            "items": {
              "$ref": "#/components/schemas/WorkflowToolRef"
            },
            "title": "Tool Set",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "workflow_id": {
            "format": "uuid",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflow_id",
          "version",
          "status",
          "prompt_template",
          "input_schema",
          "output_schema",
          "model_config",
          "policy_pack",
          "tool_set",
          "guardrails",
          "eval_dataset_bindings",
          "created_at"
        ],
        "title": "WorkflowVersionRead",
        "type": "object"
      },
      "WorkflowVersionStatus": {
        "enum": [
          "draft",
          "published",
          "archived"
        ],
        "title": "WorkflowVersionStatus",
        "type": "string"
      },
      "WorkflowVersionUpdate": {
        "properties": {
          "eval_dataset_bindings": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Dataset Bindings"
          },
          "guardrails": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardrails"
          },
          "input_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Schema"
          },
          "model_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "output_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Schema"
          },
          "policy_pack": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WorkflowPolicyRule"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Pack"
          },
          "prompt_template": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Template"
          },
          "rollout_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowRolloutConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "tool_set": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WorkflowToolRef"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Set"
          },
          "version": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "WorkflowVersionUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Agent Reliability Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/workflows": {
      "get": {
        "operationId": "list_workflows_api_v1_projects__project_id__workflows_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowRead"
                  },
                  "title": "Response List Workflows Api V1 Projects  Project Id  Workflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "create_workflow_api_v1_projects__project_id__workflows_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/v1/workflow-versions/{workflow_version_id}": {
      "get": {
        "operationId": "get_workflow_version_api_v1_workflow_versions__workflow_version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow Version",
        "tags": [
          "workflows"
        ]
      },
      "patch": {
        "operationId": "update_workflow_version_api_v1_workflow_versions__workflow_version_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowVersionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workflow Version",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflow-versions/{workflow_version_id}/publish": {
      "post": {
        "operationId": "publish_workflow_version_api_v1_workflow_versions__workflow_version_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishWorkflowVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Workflow Version",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/v1/workflows/{workflow_id}/rollout/activate": {
      "post": {
        "operationId": "activate_workflow_rollout_api_v1_workflows__workflow_id__rollout_activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateWorkflowRolloutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Workflow Rollout",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{workflow_id}/rollout/monitor": {
      "post": {
        "operationId": "monitor_workflow_rollout_api_v1_workflows__workflow_id__rollout_monitor_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolloutMonitorRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Monitor Workflow Rollout",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{workflow_id}/versions": {
      "get": {
        "operationId": "list_workflow_versions_api_v1_workflows__workflow_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowVersionRead"
                  },
                  "title": "Response List Workflow Versions Api V1 Workflows  Workflow Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workflow Versions",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "create_workflow_version_api_v1_workflows__workflow_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowVersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow Version",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}